fix: fixed bugs and adjusted requested permissions
1. Fixed a bug regarding reappearing requests after clearing them
2. Fixed a bug regarding mocking script not being injected
