+++ manifest.json
@@ -3,7 +3,7 @@
   "name": "GraphQL Network Tab",
   "version": "3.1.0",
   "devtools_page": "devtools_page.html",
-  "permissions": ["clipboardWrite", "scripting", "storage", "tabs", "webRequest"],
+  "permissions": ["scripting", "storage", "tabs"],
   "host_permissions": ["<all_urls>"],
   "content_scripts": [
     {
